api/requirements.txt
Flask
pandas
requests
schedule
datetime
python-dotenv